support ThrustMaster Pro Digital 2 wheel build
TM Pro Digital 2 wheel is a racing wheel with 2 axes, X for steering,
half of Y axis for gas, and half for brakes.  Also requires setting
wheelIndex = pedalIndex in SteeringWheelInputController, but I haven't
quite yet figured out the best way to go about doing that in an easily
configurable fashion.

This update removes the restriction of requiring multiple input devices
to be able to read pedals -- the TM, and presumably other wheels, have
the same device number for both the pedals and the wheel.