make an attempt to improve steering wheel and pedal detection
Attempt to use Windows enumeration of "driving devices" to improve
detection of what DirectInput devices should be mapped to pedal / wheel
controllers.  If that fails, fall back to previous method of assuming
that if there are 2 DirectInput devices, the one with Force Feedback
support must be the steering wheel, and the one that doesn't have FF must
by the pedals.  If there is only 1 DirectInput device, then assume that
it should be a wheel with pedals, even if it doesn't enumerate that it
has driving abilities.

If there is only one device, and it has 1 pedal, or we can't get a number
of pedals, assume that Pedals are on the Y axis -- this is a relatively
safe assumption as we already assume that the wheel is on the X axis.
Also assume that the gas pedal takes one half of the axis and the brake
takes the other half.  This maps to my old ThrustMaster, and probably
many other wheel/pedal combos.

If there is only one device, and we can reliably determine that it has 2 or
3 pedals, then assume gas is "Y" axis and brake is "Z" axis.

This will need to be tested with other wheel / pedal devices to determine
if this method works well.  Works fine with my old TM wheel, but my old
TM wheel doesn't identify as a steering device.