many adjustments to allow project to be placed inside another project
Move following out of main project:
 DirectInputPlugin, PFRTester, Tools, project_freeride

To place this project inside another, first create a new Project in Unity,
then copy the contents of this project's Assets/StreamingAssets folder to
your new project's Assets/StreamingAssets folder.  Then copy the contents
of this project's ProjectSettings folder to your new project's
ProjectSettings folder.  Copy everything that remains to your new
project's Assets folder, in it's own subfolder.

Or do the same operation, but do it with a git submodule. :)

While figuring all of this operation out, I discovered that the project
was writing data to Application.streamingAssetsPath, which is not
guaranteed to be a writeable location.  All data that was being read from
Application.streamingAssetsPath is now also read from
Application.persistentDataPath, and when changes to the data occur, it will
be written to Application.persistentDataPath.

The code updates that handle these changes could be done better, but I was
primarily concerned with how to get it working, rather than how to do it
best in a language I am not yet familiar with.

Also, fix condition that would halt game if the PlayerCar layer was not
defined (Layer definitions are contained in /ProjectSettings )